organized the tests to exapand testing of data
