+++ smrf/framework/CoreConfig.ini
@@ -195,6 +195,11 @@ default = None,
 type = CriticalDirectory,
 description = Path to the top level directory where multiple HRRR gridded dataset are located
 
+hrrr_load_method:
+default = first,
+options = [first timestep],
+description = Method to load the HRRR data either load all data first or for each timestep
+
 wrf_file:
 type =  CriticalFilename,
 description = Path to the netCDF file containing WRF data
