+++ smrf/framework/recipes.ini
@@ -327,7 +327,7 @@ trigger:
   has_value = [gridded data_type netcdf]
 
 gridded:
-  remove_item = [wrf_file hrrr_directory hrrr_forecast_flag],
+  remove_item = [wrf_file hrrr_directory hrrr_forecast_flag hrrr_load_method],
 
 [dk_recipe]
 trigger:
